// packages/src/api/generated/semantic/model.ts
export type GetSemanticTermsStatus = (typeof GetSemanticTermsStatus)[keyof typeof GetSemanticTermsStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetSemanticTermsStatus = {
  draft: 'draft',
  certified: 'certified',
  deprecated: 'deprecated',
} as const;

export type SizeParams = number;

export type PageParams = number;
// ...
export type GetSemanticTermsParams = {
  page?: PageParams;
  size?: SizeParams;
  /**
   * 搜索术语名称 / 别名
   */
  keyword?: string;
  /**
   * 状态筛选
   */
  status?: GetSemanticTermsStatus;
  /**
   * 仅看已认证
   */
  onlyCertified?: boolean;
};